renamed GlobalVariables to GlobalConstants
